Micro optimizations and safeguards (#58)
* change everything to u8 if possible - add dev panic script

* removed unneeded as u32

+++ crates/unigen/src/builder/core.rs
@@ -18,7 +18,7 @@ pub struct Block {
 
 #[derive(Debug, Copy, Clone)]
 pub struct Atom {
-    pub electrons: u32,
+    pub electrons: u8,
     pub nucleus: Nucleus,
 }
